refactor: piece post, put api 컨트롤러가 같은 Dto를 사용하도록 수정
