--- src/main/java/chess/dto/request/BoardRequestDto.java
@@ -1,27 +1,0 @@
-package chess.dto.request;
-
-public class BoardRequestDto {
-
-    private final int roomId;
-    private final String source;
-    private final String target;
-
-    public int getRoomId() {
-        return roomId;
-    }
-
-    public BoardRequestDto(int roomId, String source, String target) {
-        this.roomId = roomId;
-        this.source = source;
-        this.target = target;
-    }
-
-    public String getSource() {
-        return source;
-    }
-
-    public String getTarget() {
-        return target;
-    }
-
-}
