+++ src/main/java/chess/dto/request/PiecesRequestDto.java
@@ -2,9 +2,14 @@ package chess.dto.request;
 
 public class PiecesRequestDto {
 
-    private int roomId;
+    private final int roomId;
+    private String source;
+    private String target;
 
-    public PiecesRequestDto() {
+    public PiecesRequestDto(int roomId, String source, String target) {
+        this.roomId = roomId;
+        this.source = source;
+        this.target = target;
     }
 
     public PiecesRequestDto(int roomId) {
@@ -14,4 +19,13 @@ public class PiecesRequestDto {
     public int getRoomId() {
         return roomId;
     }
+
+    public String getSource() {
+        return source;
+    }
+
+    public String getTarget() {
+        return target;
+    }
+
 }
